Abstraction

+++ src/Kirbo/Database.cs
@@ -21,13 +21,18 @@ namespace Kirbo
 
 		public void Save()
 		{
-			foreach (var playlistPath in Directory.GetFiles(Config.playlistsPath))
-			{
-				File.Delete(playlistPath);
-			}
+			var usedPlaylistsPaths = new List<string>();
 			foreach (var playlist in playlists)
 			{
 				File.WriteAllText(Config.playlistsPath + playlist.systemName + ".json", JsonConvert.SerializeObject(playlist));
+			}
+
+			foreach (var playlistPath in Directory.GetFiles(Config.playlistsPath))
+			{
+				if (!usedPlaylistsPaths.Contains(playlistPath.CleanPath()))
+				{
+					File.Delete(playlistPath);
+				}
 			}
 		}
 
@@ -67,22 +72,61 @@ namespace Kirbo
 				playlistsNotebook.RemovePage(i);
 			}
 
+			// Generate all music
 			{
+				var allMusicList = new TreeView() { HeadersClickable = true };
+				// songsList.HeadersVisible = false;
+
+				var listStore = new ListStore(typeof(string), typeof(string), typeof(string));
+				allMusicList.Model = listStore;
+
+				var titleCellView = new CellRendererText();
+				var titleColumn = new TreeViewColumn("Title", titleCellView);
+				titleColumn.AddAttribute(titleCellView, "text", 0);
+				allMusicList.AppendColumn(titleColumn);
+
+				var artistCellView = new CellRendererText();
+				var artistColumn = new TreeViewColumn("Artist", artistCellView);
+				artistColumn.AddAttribute(artistCellView, "text", 1);
+				allMusicList.AppendColumn(artistColumn);
+
+				var albumCellView = new CellRendererText();
+				var albumColumn = new TreeViewColumn("Album", albumCellView);
+				albumColumn.AddAttribute(albumCellView, "text", 2);
+				allMusicList.AppendColumn(albumColumn);
+
+				var pathCellView = new CellRendererText();
+				var pathColumn = new TreeViewColumn("Location", albumCellView);
+				pathColumn.AddAttribute(pathCellView, "text", 2);
+				allMusicList.AppendColumn(pathColumn);
+
+				foreach (var song in MainWindow.current.database.songs)
+				{
+					listStore.AppendValues(song.title, song.artist, song.album, song.path);
+				}
+
+				playlistsNotebook.AppendPage(allMusicList, new Label("All Music"));
 			}
 
 			foreach (var playlist in playlists)
 			{
 				var page = new ListBox();
 
-				page.Add(new Entry(playlist.title));
-				page.Add(new Entry(playlist.description));
+				var hbox = new HBox(false, 8);
+				var titleInput = new Entry(playlist.title);
+				titleInput.Changed += (sender, args) => playlist.title = titleInput.Text;
+				hbox.Add(titleInput);
 				var playButton = new Button(new Label("Play"));
 				playButton.Clicked += delegate
 				{
 					MainWindow.current.player.LoadPlaylist(playlist);
 					MainWindow.current.player.PlayRandomSongFromPlaylist();
 				};
-				page.Add(playButton);
+				hbox.Add(playButton);
+				page.Add(hbox);
+				var descInput = new Entry(playlist.description);
+				descInput.Changed += (sender, args) => playlist.description = descInput.Text;
+				page.Add(descInput);
 
 				var songsList = new TreeView() { HeadersClickable = true };
 				// songsList.HeadersVisible = false;
